# Conflicts:
#	src/main/java/com/example/LuanVanTotNghiep/configuration/SecurityConfig.java

+++ src/main/java/com/example/LuanVanTotNghiep/service/GroupsService.java
@@ -23,10 +23,14 @@ public class GroupsService {
     GroupsRepository groupsRepository;
     GroupsMapper groupsMapper;
     public GroupsResponse createGroup(GroupsRequest request){
-        if(groupsRepository.existsByContent(request.getContent())){
+        if(groupsRepository.existsById(request.getGroup_id())){
+            throw new AppException(ErrorCode.GROUP_EXISTS);
+        }
+        if(groupsRepository.existsByGroupname(request.getGroupname())){
             throw  new AppException(ErrorCode.CONTENT_EXISTS);
         }
         Groups groups = groupsMapper.toCreateGroups(request);
+        //groups.setType(request.isType());
         return groupsMapper.toGroupsResponse(groupsRepository.save(groups));
     }
     public List<GroupsResponse> getAllGroups(){
@@ -34,11 +38,24 @@ public class GroupsService {
         return groupsList;
     }
     public void deleteGroups(String id){
+        if (groupsRepository.existsByGroupId(id)){
+            throw new AppException(ErrorCode.CONSTRAINT_VIOLATION);
+        }
         groupsRepository.deleteById(id);
     }
-    public GroupsResponse updateGroups(String id, GroupsRequest request){
-        Groups groups = groupsRepository.findById(id).orElseThrow(()->new AppException(ErrorCode.CONTENT_NO_EXISTS));
+    public GroupsResponse updateGroups(String group_id, GroupsRequest request){
+        Groups groups = groupsRepository.findById(group_id).orElseThrow(()->new AppException(ErrorCode.CONTENT_NO_EXISTS));
+        if (groupsRepository.existsByGroupId(group_id)){
+            throw new AppException(ErrorCode.CONSTRAINT_VIOLATION);
+        }
         groupsMapper.updateGroups(groups,request);
         return groupsMapper.toGroupsResponse(groupsRepository.save(groups));
     }
+    public List<GroupsResponse> getAllGroupsTrue(){
+        return groupsMapper.listGroups(groupsRepository.findGroupsByTypeTrue());
+    }
+//    public void typegroup(String id, boolean type){
+//        Groups groups = groupsRepository.findById(id).orElseThrow(()->new AppException(ErrorCode.CONTENT_NO_EXISTS));
+//        if ()
+//    }
 }
